# svehi-backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SVEHI Modernización API
  description: API para el sistema de Gestión de Vehículos e Históricos (SVEHI)
  version: 1.0.0
  contact:
    name: Arquitectura SVEHI
servers:
  - url: http://localhost:8080/svehi
    description: Local Server
...
paths:
...
  /v1/vehiculos:
    get:
      tags:
        - Vehiculos
      summary: Listar vehículos
      operationId: getVehiculos
      parameters:
        - name: matricula
          in: query
          schema:
            type: string
        - name: marcaId
          in: query
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehiculoVO'
    post:
      tags:
        - Vehiculos
      summary: Crear vehículo
      operationId: createVehiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehiculoVO'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiculoVO'
...
  /v1/vehiculos/{id}/matriculas:
    get:
      tags: [Vehiculos]
      summary: Listar matrículas
      operationId: getVehiculoMatriculas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatriculaVO'
    post:
      tags: [Vehiculos]
      summary: Añadir matrícula
      operationId: addVehiculoMatricula
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatriculaVO'
      responses:
        '201':
          description: Created
  /v1/vehiculos/{id}/equipamiento:
    get:
      tags: [Vehiculos]
      summary: Listar equipamiento
      operationId: getVehiculoEquipamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipamientoVO'
    post:
      tags: [Vehiculos]
      summary: Añadir equipamiento
      operationId: addVehiculoEquipamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipamientoVO'
      responses:
        '201':
          description: Created
  /v1/vehiculos/{id}/mantenimientos:
    get:
      tags: [Vehiculos]
      summary: Listar mantenimientos
      operationId: getVehiculoMantenimientos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MantenimientoVO'
    post:
      tags: [Vehiculos]
      summary: Añadir mantenimiento
      operationId: addVehiculoMantenimiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MantenimientoVO'
      responses:
        '201':
          description: Created
...
components:
  schemas:
    VehiculoVO:
      type: object
      required:
        - matricula
        - marcaId
        - modelo
      properties:
        id:
          type: integer
          format: int64
        matricula:
          type: string
        marcaId:
          type: integer
          format: int64
        marcaNombre:
          type: string
        modelo:
          type: string
        tipoVehiculo:
          type: string
        bastidor:
          type: string
        fechaMatriculacion:
          type: string
          format: date
        activo:
          type: boolean
          default: true
    PageVehiculoVO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VehiculoVO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
    MatriculaVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numero:
          type: string
        fecha:
          type: string
          format: date
        actual:
          type: boolean
    EquipamientoVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        descripcion:
          type: string
    MantenimientoVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        descripcion:
          type: string
        fecha:
          type: string
          format: date
        importe:
          type: number
          format: double
        finalizado:
          type: boolean
...